requirements.txt
Django==3.2.8
psycopg2-binary==2.9.1
pip~=23.2.1
wheel~=0.41.2
pytz~=2023.3.post1
sqlparse~=0.4.4
psycopg2~=2.9.9
asgiref~=3.7.2
setuptools~=68.2.0
Jinja2~=3.1.2
colorama~=0.4.6
djangorestframework~=3.14.0